get default list ID, update checkbox initial state with default status

+++ app/src/main/java/com/jeffcunningham/twitterlistviewer_android/restapi/dto/DefaultList.java
@@ -14,7 +14,7 @@ public class DefaultList {
     private String alias;
     @SerializedName("listId")
     @Expose
-    private Integer listId;
+    private Long listId;
     @SerializedName("__v")
     @Expose
     private Integer v;
@@ -38,11 +38,11 @@ public class DefaultList {
         this.alias = alias;
     }
 
-    public Integer getListId() {
+    public Long getListId() {
         return listId;
     }
 
-    public void setListId(Integer listId) {
+    public void setListId(Long listId) {
         this.listId = listId;
     }
